chore: refactor clients
