+++ pyproject.toml
@@ -16,6 +16,7 @@ classifiers = [
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10"
 ]
+repository = "https://github.com/numaproj/numalogic-prometheus"
 
 [tool.poetry.dependencies]
 python = ">=3.9, <3.11"
