+++ tests/__init__.py
@@ -5,7 +5,7 @@ import fakeredis
 server = fakeredis.FakeServer()
 redis_client = fakeredis.FakeStrictRedis(server=server, decode_responses=True)
 
-with patch("numaprom.redis.get_redis_client") as mock_get_redis_client:
+with patch("numaprom.clients.redis.get_redis_client") as mock_get_redis_client:
     mock_get_redis_client.return_value = redis_client
     from numaprom.udf import window
     from numaprom.udsink import train, train_rollout
